+++ examples/account/address/outputs.tf
@@ -1,0 +1,4 @@
+output "address" {
+  description = "The Generated Address"
+  value       = twilio_account_address.address
+}