+++ twilio/internal/services/account/data_source_account_address.go
@@ -1,0 +1,122 @@
+package account
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/RJPearson94/terraform-provider-twilio/twilio/common"
+	"github.com/RJPearson94/terraform-provider-twilio/twilio/utils"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+)
+
+func dataSourceAccountAddress() *schema.Resource {
+	return &schema.Resource{
+		Read: dataSourceAccountAddressRead,
+
+		Timeouts: &schema.ResourceTimeout{
+			Read: schema.DefaultTimeout(5 * time.Minute),
+		},
+
+		Schema: map[string]*schema.Schema{
+			"sid": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"account_sid": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"customer_name": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"friendly_name": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"street": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"street_secondary": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"city": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"region": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"postal_code": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"iso_country": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"emergency_enabled": {
+				Type:     schema.TypeBool,
+				Computed: true,
+			},
+			"validated": {
+				Type:     schema.TypeBool,
+				Computed: true,
+			},
+			"verified": {
+				Type:     schema.TypeBool,
+				Computed: true,
+			},
+			"date_created": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"date_updated": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+		},
+	}
+}
+
+func dataSourceAccountAddressRead(d *schema.ResourceData, meta interface{}) error {
+	client := meta.(*common.TwilioClient).API
+	ctx, cancel := context.WithTimeout(meta.(*common.TwilioClient).StopContext, d.Timeout(schema.TimeoutRead))
+	defer cancel()
+
+	getResponse, err := client.Account(d.Get("account_sid").(string)).Address(d.Get("sid").(string)).FetchWithContext(ctx)
+	if err != nil {
+		if utils.IsNotFoundError(err) {
+			d.SetId("")
+			return nil
+		}
+		// If the account sid is incorrect a 401 is returned, a this is a generic error this will not be handled here and an error will be returned
+		return fmt.Errorf("[ERROR] Failed to read address: %s", err.Error())
+	}
+
+	d.SetId(getResponse.Sid)
+	d.Set("sid", getResponse.Sid)
+	d.Set("account_sid", getResponse.AccountSid)
+	d.Set("customer_name", getResponse.CustomerName)
+	d.Set("friendly_name", getResponse.FriendlyName)
+	d.Set("street", getResponse.Street)
+	d.Set("street_secondary", getResponse.StreetSecondary)
+	d.Set("city", getResponse.City)
+	d.Set("region", getResponse.Region)
+	d.Set("postal_code", getResponse.PostalCode)
+	d.Set("iso_country", getResponse.IsoCountry)
+	d.Set("emergency_enabled", getResponse.EmergencyEnabled)
+	d.Set("validated", getResponse.Validated)
+	d.Set("verified", getResponse.Verified)
+	d.Set("date_created", getResponse.DateCreated.Time.Format(time.RFC3339))
+
+	if getResponse.DateUpdated != nil {
+		d.Set("date_updated", getResponse.DateUpdated.Time.Format(time.RFC3339))
+	}
+
+	return nil
+}