+++ twilio/internal/services/account/registration.go
@@ -12,8 +12,10 @@ func (r Registration) Name() string {
 // SupportedDataSources returns the supported Data Sources supported by this Service
 func (r Registration) SupportedDataSources() map[string]*schema.Resource {
 	return map[string]*schema.Resource{
-		"twilio_account_balance": dataSourceAccountBalance(),
-		"twilio_account_details": dataSourceAccountDetails(),
+		"twilio_account_balance":   dataSourceAccountBalance(),
+		"twilio_account_details":   dataSourceAccountDetails(),
+		"twilio_account_address":   dataSourceAccountAddress(),
+		"twilio_account_addresses": dataSourceAccountAddresses(),
 	}
 }
 
@@ -21,5 +23,6 @@ func (r Registration) SupportedDataSources() map[string]*schema.Resource {
 func (r Registration) SupportedResources() map[string]*schema.Resource {
 	return map[string]*schema.Resource{
 		"twilio_account_sub_account": resourceAccountSubAccount(),
+		"twilio_account_address":     resourceAccountAddress(),
 	}
 }